refactor(file): rename module file
